# docs/api.yml
openapi: 3.0.0
servers:
  - url: https://www.bhut.com.br
info:
  version: 1.0.0
  title: TIMESHEET API
  description: API for timesheet
  contact:
    name: BHUT
    url: 'https://www.bhut.com.br'
    email: 'contato@bhut.com.br'
  license:
    name: MIT
    url: https://api.github.com/licenses/mit
...
paths:
...
  /v1/timesheet/{year}/{month}:
    get:
      tags:
        - Timesheet
      summary: 'Get timesheet report'
      description: 'Get an existent timesheet report'
      operationId: getTimesheetReport
      parameters:
        - name: year
          in: path
          description: 'Year to filter'
          schema:
            type: integer
            example: 2024
          required: true
        - name: month
          in: path
          description: 'Month to filter'
          schema:
            type: integer
            example: 08
          required: true
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetResponse'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
  /v1/timesheet/{year}/{month}/{day}:
    post:
      tags:
        - Timesheet
      summary: 'Update timesheet report'
      description: 'Update an existent timesheet report'
      operationId: postTimesheetReport
      parameters:
        - name: year
          in: path
          description: 'Year to filter'
          schema:
            type: integer
            example: 2024
          required: true
        - name: month
          in: path
          description: 'Month to filter'
          schema:
            type: integer
            example: 08
          required: true
        - name: day
          in: path
          description: 'Day to filter'
          schema:
            type: integer
            example: 25
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimesheetRequest'
        required: true
      responses:
        '204':
          description: 'Request processed'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
  # USER ENDPOINTS
  /v1/user:
    post:
      tags:
        - User
      summary: 'Create User'
      description: 'Insert a user into the api'
      operationId: createUser
      requestBody:
        description: 'User data'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
    get:
      tags:
        - User
      summary: 'List user'
      description: 'Return an user list'
      operationId: listUser
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ListUserResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
  /v1/user/{id}:
    get:
      tags:
        - User
      summary: 'Get user'
      description: 'Get user details'
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: 'Identifier to filter'
          schema:
            type: string
            format: uuid
            example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
          required: true
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
    put:
      tags:
        - User
      summary: 'Update user'
      description: 'Update an existent user'
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: 'Identifier to filter'
          schema:
            type: string
            format: uuid
            example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
          required: true
      requestBody:
        description: 'User data'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '204':
          description: 'Request processed'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
...
  /v1/user/{id}/report/{year}/{month}:
    get:
      tags:
        - User
      summary: 'Get user report'
      description: 'Get an existent user report'
      operationId: getUserReport
      parameters:
        - name: id
          in: path
          description: 'Identifier to filter'
          schema:
            type: string
            format: uuid
            example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
          required: true
        - name: year
          in: path
          description: 'Year to filter'
          schema:
            type: integer
            example: 2024
          required: true
        - name: month
          in: path
          description: 'Month to filter'
          schema:
            type: integer
            example: 08
          required: true
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReportResponse'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
  /v1/user/{id}/report/{year}/{month}/closed:
    patch:
      tags:
        - User
      summary: 'Close user report'
      description: 'Close user report'
      operationId: closeUserReport
      parameters:
        - name: id
          in: path
          description: 'Identifier to filter'
          schema:
            type: string
            format: uuid
            example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
          required: true
        - name: year
          in: path
          description: 'Year to filter'
          schema:
            type: integer
            example: 2024
          required: true
        - name: month
          in: path
          description: 'Month to filter'
          schema:
            type: integer
            example: 08
          required: true
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloseUserReportResponse'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
components:
  schemas:
...
    CloseUserReportResponse:
      type: object
      properties:
        name:
          type: string
          description: 'User name'
          example: 'John Doe'
        team:
          type: string
          description: 'User team'
          example: 'Bhut team'
        bankHours:
          type: boolean
          description: 'Defines if user has bank hours'
          example: true
        calendar:
          type: object
          description: 'Calendar data'
          allOf:
            - $ref: '#/components/schemas/ReportCalendarResponse'
        pay:
          type: object
          description: 'Pay data'
          allOf:
            - $ref: '#/components/schemas/ReportPayResponse'
...
    CreateTimesheetRequest:
      type: object
      required:
        - period
      properties:
        period:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/CreateTimesheetPeriodRequest'
    CreateTimesheetPeriodRequest:
      type: object
      required:
        - start
        - end
        - description
      properties:
        start:
          type: string
          description: 'Start of the day'
          example: '08:30'
        end:
          type: string
          description: 'End of the day'
          example: '17:30'
        description:
          type: string
          description: 'Description of the day'
          example: 'Implement feature BHUT-2001'
...
    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - password
        - team
        - hourValue
        - hasBankHours
        - contractTotal
        - groupId
        - startDate
      properties:
        name:
          type: string
          description: 'User name'
          example: 'John Doe'
        email:
          type: string
          description: 'User email'
          example: 'email@domain.com'
        password:
          type: string
          description: 'User password'
          example: 'password@123'
        team:
          type: string
          description: 'User team'
          example: 'Bhut team'
        hourValue:
          type: number
          description: 'User hour value'
          example: 17.05
        hasBankHours:
          type: boolean
          description: 'Define if user has bank hours'
          example: true
        contractTotal:
          type: integer
          description: 'User contract total hours'
          example: 176
        groupId:
          type: string
          description: 'User group identifier'
          example: '6de140ca-ae94-4e1d-9c32-8de18fa22e45'
        startDate:
          type: string
          description: 'User start date'
          example: '2022-01-04'
    CreateUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 'User identifier'
          example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
...
    ListUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 'User identifier'
          example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
        name:
          type: string
          description: 'User name'
          example: 'John Doe'
        email:
          type: string
          description: 'User email'
          example: 'email@domain.com'
        team:
          type: string
          description: 'User team'
          example: 'Bhut team'
        hourValue:
          type: number
          description: 'User hour value'
          example: 17.05
        hasBankHours:
          type: boolean
          description: 'Define if user has bank hours'
          example: true
        totalBankHours:
          type: number
          description: 'User total bank hours'
          example: 9.51
        contractTotal:
          type: integer
          description: 'User contract total hours'
          example: 176
        groupId:
          type: string
          description: 'User group identifier'
          example: '6de140ca-ae94-4e1d-9c32-8de18fa22e45'
        groupName:
          type: string
          description: 'User group name'
          example: 'Developer'
...
    ReportCalendarResponse:
      type: object
      properties:
        month:
          type: string
          description: 'Report month'
          example: '08'
        year:
          type: string
          description: 'Report year'
          example: '2024'
        total:
          type: string
          description: 'Report total hours'
          example: '176'
    ReportDaysResponse:
      type: object
      properties:
        date:
          type: string
          description: 'Report days date'
          example: '2024-08-30'
        businessDay:
          type: boolean
          description: 'Defines whether it is a business day or not'
          example: true
        period:
          type: array
          description: 'Period data'
          items:
            allOf:
              - $ref: '#/components/schemas/ReportPeriodResponse'
        total:
          type: string
          description: 'Total worked hours of the day'
          example: '08:00'
    ReportPayResponse:
      type: object
      properties:
        date:
          type: string
          description: 'Report pay date'
          example: '2024-08-30'
        hour:
          type: number
          description: 'Report hour date'
          example: 10
        total:
          type: string
          description: 'Total hours'
          example: '176'
    ReportPeriodResponse:
      type: object
      properties:
        start:
          type: string
          description: 'Start of the day'
          example: '08:30'
        end:
          type: string
          description: 'End of the day'
          example: '17:30'
        description:
          type: string
          description: 'Description of the day'
          example: 'Implement feature BHUT-2001'
...
    ResourceNotFoundResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                enum:
                  - 'RESOURCE_NOT_FOUND'
              message:
                type: string
                example: 'Resource not found'
    TimesheetResponse:
      type: object
      properties:
        closed:
          type: boolean
          description: 'Defines if the report is closed or not'
          example: true
        month:
          type: integer
          description: 'Report month'
          example: 08
        year:
          type: integer
          description: 'Report year'
          example: 2024
        total:
          type: string
          description: 'Sum of report hours'
          example: '176'
        balance:
          type: number
          description: 'Hour balance'
          example: 10.30
        days:
          type: array
          description: 'Days data'
          items:
            allOf:
              - $ref: '#/components/schemas/ReportDaysResponse'
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: 'User name'
          example: 'John Doe'
        email:
          type: string
          description: 'User email'
          example: 'email@domain.com'
        team:
          type: string
          description: 'User team'
          example: 'Bhut team'
        hourValue:
          type: number
          description: 'User hour value'
          example: 17.05
        hasBankHours:
          type: boolean
          description: 'Define if user has bank hours'
        contractTotal:
          type: integer
          description: 'User contract total hours'
          example: 176
        groupId:
          type: string
          description: 'User group identifier'
          example: '6de140ca-ae94-4e1d-9c32-8de18fa22e45'
        startDate:
          type: string
          description: 'User start date'
          example: '2022-01-04'
    UserReportResponse:
      type: object
      properties:
        closed:
          type: boolean
          description: 'Defines if the report is closed or not'
          example: true
        month:
          type: integer
          description: 'Report month'
          example: 08
        year:
          type: integer
          description: 'Report year'
          example: 2024
        total:
          type: string
          description: 'Sum of report hours'
          example: '176'
        balance:
          type: number
          description: 'Hour balance'
          example: 10.30
        days:
          type: array
          description: 'Days data'
          items:
            allOf:
              - $ref: '#/components/schemas/ReportDaysResponse'
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 'User identifier'
          example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
        name:
          type: string
          description: 'User name'
          example: 'John Doe'
        email:
          type: string
          description: 'User email'
          example: 'email@domain.com'
        team:
          type: string
          description: 'User team'
          example: 'Bhut team'
        hourValue:
          type: number
          description: 'User hour value'
          example: 17.05
        hasBankHours:
          type: boolean
          description: 'Define if user has bank hours'
          example: true
        totalBankHours:
          type: number
          description: 'User total bank hours'
          example: 9.51
        contractTotal:
          type: integer
          description: 'User contract total hours'
          example: 176
        groupId:
          type: string
          description: 'User group identifier'
          example: '6de140ca-ae94-4e1d-9c32-8de18fa22e45'
        groupName:
          type: string
          description: 'User group name'
          example: 'Developer'
        startDate:
          type: string
          description: 'User start date'
          example: '2022-01-04'
    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                enum:
                  - 'INVALID_FIELD'
                  - 'INVALID_HEADER'
              message:
                type: string
                example: 'Invalid/required fields'
  securitySchemes:
    JwtAuth:
      type: controllers
      scheme: bearer
      bearerFormat: JWT
    ApiAuth:
      in: header
      type: apiKey
      name: X-Api-Key